refactor(init): extract creation flow (#520)
